requirements.txt
boto3
crowdstrike-falconpy
google-cloud-securitycenter
google-cloud-resource-manager >= 1.0.2
tls-syslog
google-auth
google-api-python-client
py7zr
